git-svn-id: http://i-path.goto.info.waseda.ac.jp/svn/i-Path/trunk@102 148987d4-aee6-11dd-a2c1-ab1b71097782

+++ SIRENSJNI/SIRENSSocket.java
@@ -38,7 +38,7 @@ import java.net.*;
 
 public class SIRENSSocket extends Socket {
     public native void setSockoptSIRENSIDX (int srsmax, int[] mode, int[] probe, int[] qttlmax, int[] qttlmin, int[] sttlmax, int[] sttlmin) throws IOException;
-    public native void getSockoptSIRENSSDATA(int dir, int mode, int probe, int [] array) throws IOException;
+    public native void getSockoptSIRENSSDATA(int dir, int mode, int probe, long [] array) throws IOException;
     static {
         System.loadLibrary("SIRENSImpl");
     }
